OSX Gitian .yml
  * updated macdeploy
  * updated shadow.pro to handle OSX cross compile
  * updated qt-osx.yml to handle build properly

At the moment the build is not deterministic and signing hasn't yet been completed.
We still need to add HDI utils to create the .dmg file, as well as update our images to look more Shadow like :)

We need to implement versioning, signing and automatic upload to web server, but we'll get there soon!